# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "description": "Example info",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://example.localhost",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/troops/{id}/patrol": {
            "get": {
                "tags": [
                    "Troops",
                    "Patrols"
                ],
                "summary": "Get all patrols of a troop",
                "description": "Returns all patrols belonging to a given troop.",
                "operationId": "9260105616f6cabd503b4064bb968480",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of patrols"
                    },
                    "404": {
                        "description": "Troop not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Troops",
                    "Patrols"
                ],
                "summary": "Create a new patrol in a troop",
                "description": "Creates a new patrol (gang) within the specified troop.",
                "operationId": "5be23e51f178fb3884a7525677ac8986",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the troop to add the patrol to",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data of the new patrol",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Wolves"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#1a73e8"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Patrol successfully created"
                    },
                    "400": {
                        "description": "Missing name"
                    },
                    "404": {
                        "description": "Troop not found"
                    }
                }
            }
        },
        "/troops/{id}/members": {
            "get": {
                "tags": [
                    "Troops",
                    "Patrols"
                ],
                "summary": "Get all patrol members in a troop",
                "description": "Returns all patrol members (gang members) within a given troop.",
                "operationId": "acdd0b0ed57cb674101ca7666bc64f1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of patrol members"
                    },
                    "500": {
                        "description": "Error fetching data"
                    }
                }
            }
        },
        "/troops/{id_troop}/members/{id_user}/troop-leaders": {
            "post": {
                "tags": [
                    "Troops"
                ],
                "summary": "Assign a user as a troop leader",
                "description": "Assigns a user as a troop leader to a specific troop.",
                "operationId": "7eff843e9466c8d0a0386537bd828a67",
                "parameters": [
                    {
                        "name": "id_troop",
                        "in": "path",
                        "description": "ID of the troop",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "ID of the user to assign",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Troop leader assigned successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Database error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Troops"
                ],
                "summary": "Remove a troop leader",
                "description": "Removes a troop leader by their ID.",
                "operationId": "5b877ab31eda484489bf3e5b961134c3",
                "parameters": [
                    {
                        "name": "id_patrol_leader",
                        "in": "path",
                        "description": "ID of the troop leader to remove",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Troop leader removed"
                    },
                    "404": {
                        "description": "Leader not found"
                    },
                    "500": {
                        "description": "Database error"
                    }
                }
            }
        },
...
    },
...
}